refactor(graph-gateway): move network indexer processing to internal module
