+++ graph-gateway/tests/it_network_subgraph_client.rs
@@ -18,23 +18,23 @@ fn test_auth_token() -> String {
     std::env::var("IT_TEST_ARBITRUM_GATEWAY_AUTH").expect("Missing IT_TEST_ARBITRUM_GATEWAY_AUTH")
 }
 
-/// Test helper to build the subgraph url with the given subgraph ID.
-fn url_with_subgraph_id(name: impl AsRef<str>) -> Url {
+/// Test helper to build the subgraph url with the given deployment ID.
+fn url_with_deployment_id(name: impl AsRef<str>) -> Url {
     test_base_url()
-        .join(&format!("api/subgraphs/id/{}", name.as_ref()))
+        .join(&format!("api/deployments/id/{}", name.as_ref()))
         .expect("Invalid URL")
 }
 
 /// The Graph Network Arbitrum in the network.
 ///
 /// https://thegraph.com/explorer/subgraphs/DZz4kDTdmzWLWsV373w2bSmoar3umKKH9y82SUKr5qmp
-const GRAPH_NETWORK_ARBITRUM_SUBGRAPH_ID: &str = "DZz4kDTdmzWLWsV373w2bSmoar3umKKH9y82SUKr5qmp";
+const GRAPH_NETWORK_ARBITRUM_DEPLOYMENT_ID: &str = "QmZtNN8NbxjJ1KD5uKBYa7Gj29CT8xypSXnAmXbrLNTQgX";
 
 #[test_with::env(IT_TEST_ARBITRUM_GATEWAY_URL, IT_TEST_ARBITRUM_GATEWAY_AUTH)]
 #[tokio::test]
 async fn fetch_indexers_and_deserialize() {
     //* Given
-    let subgraph_url = url_with_subgraph_id(GRAPH_NETWORK_ARBITRUM_SUBGRAPH_ID);
+    let subgraph_url = url_with_deployment_id(GRAPH_NETWORK_ARBITRUM_DEPLOYMENT_ID);
     let auth_token = test_auth_token();
 
     let network_subgraph_client = {
@@ -63,7 +63,7 @@ async fn fetch_indexers_and_deserialize() {
 #[tokio::test]
 async fn fetch_subgraphs_and_deserialize() {
     //* Given
-    let subgraph_url = url_with_subgraph_id(GRAPH_NETWORK_ARBITRUM_SUBGRAPH_ID);
+    let subgraph_url = url_with_deployment_id(GRAPH_NETWORK_ARBITRUM_DEPLOYMENT_ID);
     let auth_token = test_auth_token();
 
     let network_subgraph_client = {
@@ -92,7 +92,7 @@ async fn fetch_subgraphs_and_deserialize() {
 #[tokio::test]
 async fn fetch_subgraph_no_l2_transfer_support_and_deserialize() {
     //* Given
-    let subgraph_url = url_with_subgraph_id(GRAPH_NETWORK_ARBITRUM_SUBGRAPH_ID);
+    let subgraph_url = url_with_deployment_id(GRAPH_NETWORK_ARBITRUM_DEPLOYMENT_ID);
     let auth_token = test_auth_token();
 
     let network_subgraph_client = {
